Requested interval's length checked against total duration of configuration.
Other than that, the internal structure of the function (accept_appointment_request)
has been changed a bit, but the API remained the same. Some refactoring done too,
and now the durations seem to be stored as a number instead of a json object,
that's making some fuss. Runtime error caused by a bug, let's say.

 On branch dev
 Your branch is up to date with 'origin/dev'.

 Changes to be committed:
	modified:   appointy/src/request_handling.cc
	modified:   tests/test.cc