Bugs hunted down around accept_appointment_request and the JSON_Parser
Everything is a big mess in my mind rn, so making this quick:
The service is now gotten from the database and immediately parsed
to an in memory Service object. The JSON_Parser was messed up
around the ids of the answer signatures (again... I don't get it).
The accept_appointment_request is still buggy: if there is plenty
time after the last booked appointment and the requested interval's
end, only an interval as big as the completion time of the
configuration is created as an offer, not the whole remaining time.

The service_to_bson functions is a miracle:
The dump() function of a json object returns a string
suitable for printing to the console, that means, the double quotes
are escaped, yet the bsoncxx::from_json() function eats it up
happily and returns the right object. I dont' undersand that either.

 On branch dev
 Your branch is up to date with 'origin/dev'.

 Changes to be committed:
	new file:   appointment_request_test3.json
	new file:   appointment_test2.json
	modified:   appointy/src/request_handling.cc
	modified:   appointy_base/include/util/json_parser.h
	modified:   appointy_base/src/util/json_parser.cc
	new file:   offers.json
	modified:   tests/test.cc