Corrected the bug mentioned in the last commit and the to_json method of some classes.
The to_json of the answer signature classes named the id as "id" instead of "_id".
Hopefully that's the last of these kind of bugs.

The service_to_bson functions miracle could be made less magical but more solid:
if an nlohmann::json object is supplied to the operator<< of a file stream,
it looks very good in the file. So I'm thinkin', it would behave alike if
it would be supplied to the operator<< of a string stream, but it may not.

 On branch dev
 Your branch is up to date with 'origin/dev'.

 Changes to be committed:
	modified:   appointy/include/db_utils.h
	modified:   appointy/include/io_ops.h
	modified:   appointy/include/request_handling.h
	modified:   appointy/src/db_utils.cc
	modified:   appointy/src/io_ops.cc
	modified:   appointy/src/request_handling.cc
	new file:   appointy_base/include/appointment/duration_request.h
	modified:   appointy_base/include/service/numeric_answer_signature.h
	new file:   appointy_base/src/appointment/duration_request.cc
	modified:   appointy_base/src/service/choice_answer_signature.cc
	modified:   loaded_services.json
	modified:   tests/premake5.lua
	modified:   tests/test.cc